+++ src/beanz_init.c
@@ -1,8 +1,0 @@
-#include <stdlib.h> // for NULL
-#include <R_ext/Rdynload.h>
-
-void R_init_idem(DllInfo *dll)
-{
-  R_registerRoutines(dll, NULL, NULL, NULL, NULL);
-  R_useDynamicSymbols(dll, TRUE);
-}